Rename files from ..graphql/pundit.. to ..graphql-pundit..
