+++ graphql-pundit.gemspec
@@ -3,7 +3,7 @@
 
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'graphql/pundit/version'
+require 'graphql-pundit/version'
 
 Gem::Specification.new do |spec|
   spec.name          = 'graphql-pundit'
@@ -28,6 +28,6 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency 'pry', '~> 0.10.4'
   spec.add_development_dependency 'bundler', '~> 1.14'
-  spec.add_development_dependency 'rake', '~> 10.0'
-  spec.add_development_dependency 'rspec', '~> 3.0'
+  spec.add_development_dependency 'rake', '~> 12.0'
+  spec.add_development_dependency 'rspec', '~> 3.6'
 end
