+++ lib/graphql-pundit/instrumenter.rb
