+++ spec/graphql-pundit/instrumenter_spec.rb
@@ -1,4 +1,4 @@
-require_relative "../../spec_helper"
+require "spec_helper"
 
 class Test
   def initialize(value)
